1_personal_blog/requirements.txt
pydantic
fastapi
python-dotenv
mongoengine